Transform scenes to scripts (and some other stuff)
Most scenes are created as scripts, mostly to get the transitions to work.
Align automations and ui-lovelace to the scene -> script change.
Replace' ios' integration with 'mobile_app'.
Add condition for 'mini-media-player' card for Sonos kitchen to be displayed.
Volume of Sonos bedroom speaker is set to 10% if it is being ungrouped.
Added automations to toggle 'opening bathroom door' automation.